Initial support for drawing a word
Add initial support for drawing a word, using the ImageDraw library.

This supports only some consonants for now, and the drawing is somewhat
simplistic.  For a word with N letters, a circle is drawn, and letters are
drawn evenly spaced around the circle.  Only the circle part of a letter
is drawn -- the lines and dots are excluded for now.

For the half-circles ("t", "sh", "r", etc), the arc is calculating by
first calculating the intersection of two circles.  The math is shamelessly
stolen from Mathworld's article on circle-circle intersections:

http://mathworld.wolfram.com/Circle-CircleIntersection.html

Signed-off-by: Timur Tabi <timur@tabi.org>